+++ src/Hand.cpp
@@ -165,7 +165,7 @@ Returned      Output stream
 *******************************************************************************/
 std::ostream& operator<< (std::ostream& rcOut, const Hand& rcHand) {
   bool bFirst = true;
-  for (int i = 0; i < rcHand.size (); i++) {
+  for (size_t i = 0; i < rcHand.size (); i++) {
     if (bFirst) {
       bFirst = false;
     }
